chore: moved the analysis methods to a dedicated folder (#2)
The methods are Tool and Diff and have been moved to the methods/ folder

This will make it easier to add new methods in the future
