+++ include/arkanjo/methods/method.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <vector>
+
+class IMethod {
+public:
+    virtual ~IMethod() = default;
+
+    virtual void execute() = 0;
+};
