+++ src/methods/diff/diff_method.cpp
@@ -1,11 +1,11 @@
 #include <arkanjo/formatter/format_manager.hpp>
-#include "duplication_finder_diff.hpp"
+#include <arkanjo/methods/diff/diff_method.hpp>
 #include <arkanjo/base/config.hpp>
 #include <arkanjo/utils/utils.hpp>
 
 using fm = FormatterManager;
 
-DuplicationFinderDiff::DuplicationFinderDiff(const fs::path& base_path_, double similarity_) {
+DiffMethod::DiffMethod(const fs::path& base_path_, double similarity_) {
     base_path = base_path_;
     similarity = similarity_;
 
@@ -14,7 +14,7 @@ DuplicationFinderDiff::DuplicationFinderDiff(const fs::path& base_path_, double 
     }
 }
 
-std::vector<std::string> DuplicationFinderDiff::find_files(const fs::path& folder_path) {
+std::vector<std::string> DiffMethod::find_files(const fs::path& folder_path) {
     std::vector<std::string> file_paths;
     for (const auto& dirEntry : std::filesystem::recursive_directory_iterator(folder_path)) {
         fs::path file_path = dirEntry.path();
@@ -25,7 +25,7 @@ std::vector<std::string> DuplicationFinderDiff::find_files(const fs::path& folde
     return file_paths;
 }
 
-bool DuplicationFinderDiff::is_empty_line(std::string line) {
+bool DiffMethod::is_empty_line(std::string line) {
     size_t line_size = line.size();
     bool is_empty = true;
     for (size_t i = 1; i < line_size; i++) {
@@ -34,7 +34,7 @@ bool DuplicationFinderDiff::is_empty_line(std::string line) {
     return is_empty;
 }
 
-std::vector<std::string> DuplicationFinderDiff::remove_blank_lines(std::vector<std::string> content) {
+std::vector<std::string> DiffMethod::remove_blank_lines(std::vector<std::string> content) {
     std::vector<std::string> ret;
     for (auto line : content) {
         if (!is_empty_line(line)) {
@@ -44,7 +44,7 @@ std::vector<std::string> DuplicationFinderDiff::remove_blank_lines(std::vector<s
     return ret;
 }
 
-bool DuplicationFinderDiff::is_equal_files(std::vector<std::string> content1, std::vector<std::string> content2) {
+bool DiffMethod::is_equal_files(std::vector<std::string> content1, std::vector<std::string> content2) {
     if (content1.size() != content2.size()) {
         return false;
     }
@@ -57,7 +57,7 @@ bool DuplicationFinderDiff::is_equal_files(std::vector<std::string> content1, st
     return true;
 }
 
-double DuplicationFinderDiff::find_similarity(std::string path1, std::string path2) {
+double DiffMethod::find_similarity(std::string path1, std::string path2) {
     std::vector<std::string> content1 = Utils::read_file_with_vector(path1);
     std::vector<std::string> content2 = Utils::read_file_with_vector(path2);
 
@@ -106,7 +106,7 @@ double DuplicationFinderDiff::find_similarity(std::string path1, std::string pat
     return similarity_metric;
 }
 
-std::vector<std::tuple<double, std::string, std::string>> DuplicationFinderDiff::find_similar_pairs(std::vector<std::string>& file_paths) {
+std::vector<std::tuple<double, std::string, std::string>> DiffMethod::find_similar_pairs(std::vector<std::string>& file_paths) {
     size_t number_files = file_paths.size();
     std::vector<std::tuple<double, std::string, std::string>> ret;
     for (size_t i = 0; i < number_files; i++) {
@@ -124,7 +124,7 @@ std::vector<std::tuple<double, std::string, std::string>> DuplicationFinderDiff:
     return ret;
 }
 
-void DuplicationFinderDiff::save_duplications(std::vector<std::tuple<double, std::string, std::string>>& file_duplication_pairs) {
+void DiffMethod::save_duplications(std::vector<std::tuple<double, std::string, std::string>>& file_duplication_pairs) {
     std::string output_file_path = base_path / "output_parsed.txt";
 
     auto fout = std::ofstream(output_file_path);
@@ -138,7 +138,7 @@ void DuplicationFinderDiff::save_duplications(std::vector<std::tuple<double, std
     fout.close();
 }
 
-void DuplicationFinderDiff::execute() {
+void DiffMethod::execute() {
     std::vector<std::string> file_paths = find_files(base_path / Config::config().source_path);
 
     std::vector<std::tuple<double, std::string, std::string>> file_duplication_pairs = find_similar_pairs(file_paths);
