+++ src/methods/tool/parser.cpp
@@ -1,4 +1,4 @@
-#include "parser.hpp"
+#include <arkanjo/methods/tool/parser.hpp>
 #include <iomanip>
 #include <iostream>
